[PassBuilderPipelines] Add a flag and 3 additional runs of the function merger
Summary:
With Oz, the inliner captures some extra size improvements from running
the function merger before it. My guess is that this occurs due to
InlineCost.cpp greatly increasing the budget for single callee
functions. The early mergers causes sets of functions that was once many
to become one and thus changing multiple-callees to single-callees and
thus they get inlined and simplified and the original function bodies
get deleted.

The post-inliner merger is added because evidently enough changes
between then and the later merger that ~20kb is saved for fb4a.

Test Plan: BSBs with MSDK bump

Reviewers: smeenai, #android_llvm

Reviewed By: smeenai

Differential Revision: https://phabricator.intern.facebook.com/D45834892