[CIR][Lowering][Bufix] Cast cir.shift amount to same width as the value
The new cir.shift op allows for distinct value and amount types, however
LLVM does not. Before this patch, the amount was not cast to the same
width as the value, breaking the lowering process. This patch fixes the
issue by casting the amount to the same width as the value.