[CIR][CIRGen] Fix calling a function through a function pointer (#467)
CIR codegen always casts the no-proto function pointer to `FuncOp`. But
the function pointer may be result of cir operations (f.e. `cir.load`).
As a result in such cases the function pointer sets to `nullptr`. That
leads to compilation error.
So this PR removes the unecessary cast to 'FuncOp' and resolves the
issue.