[CIR][IR][NFC] Fix CallOp builder with void return (#629)
One of the builders was adding a retun value to the CallOp when given a
void return type. The expected behavior is to not add a return value.
Two other minor fixes were added to the return value: its constraint was
replaced from variadic to optional and it was assigned a name. This
prevents function calls with multiple returns and facilitates access to
the single return value, respectively.