[CIR][NFC] Fix mismatch of argument type in IR tests (#837)
There is a typo in `call.cir` that uses a wrong function argument type,
leading to failure in the final LLVM IR translation.

CIR verification does not reject it, because it skips indirect calls at
the beginning. It's `verifySymbolUses` after all.


https://github.com/llvm/clangir/blob/bde154cf1243cc4f938339c4dc15b1576d3025ab/clang/lib/CIR/Dialect/IR/CIRDialect.cpp#L2672-L2679

The typo was copied to another IR test. Here we fix them all.