[CIR][CIRGen][Builtin] Allow CIRGen for builtin calls with math errorno override (#893)
As title. 
The test case used is abort(), but it is from the real code. 
Notice: Since CIR implementation for NoReturn Call is pending to
implement, the generated llvm code is like:
`define dso_local void @test() #1  {
  call void @abort(), !dbg !8
  ret void
}`
which is not right, right code should be like, 
`
`define dso_local void @test() #1  {
  call void @abort(), !dbg !8
  unreachable
}`
`
Still send this PR as Noreturn implementation is a separate issue.