[ClangIR][CIRGen] Introduce CaseOp and refactor SwitchOp (#1006)
Close https://github.com/llvm/clangir/issues/522

This solves the issue we can't handle `case` in nested scopes and we
can't handle if the switch body is not a compound statement.

The core idea of the patch is to introduce the `cir.case` operation to
the language. Then we can get the cases by traversing the body of the
`cir.switch` operation easily instead of counting the regions and the
attributes.

Every `cir.case` operation has a region and now the `cir.switch` has
only one region too. But to make the analysis and optimizations easier,
I add a new concept `simple form` here. That a simple `cir.switch`
operation is: all the `cir.case` operation owned by the `cir.switch`
lives in the top level blocks of the `cir.switch` region and there is no
other operations except the ending `cir.yield`. This solves the previous
`simplified for common-case` vs `general solution` discussion in
https://github.com/llvm/clangir/issues/522. After implemented this, I
feel the correct answer to it is, we want a general solution for
constructing and lowering the operations but we like simple and common
case for analysis and optimizations. We just mixed the different phases.

For other semantics, see `CIROps.td`.

For lowering, we can make it generally by lower the cases one by one and
finally lower the switch itself.

Although this patch has 1000+ lines of changes, I feel it is relatively
neat especially it erases some odd behaviors before.

Tested with Spec2017's C benchmarks except 500.perlbench_r.