[CIR] Lower nested local constant alloca (#1261)
This PR adds support for lowering local constants in nested scopes,
including those in nested loops.

For those constant allocas in non-loop inner scopes, this PR keeps their
constant flags during alloca hoisting. LLVM lowering would correctly
emit necessary invariant metadata for those allocas.

For those constant allocas in a loop, this PR introduces a new operation
`cir.invariant_group` that marks the beginning of the lifetime of the
constant objects. This operation is put at the location of the alloca
operation before hoisting them. This PR updates LLVM lowering to emit
the necessary invariant metadata when loading and storing through such
pointers.

This PR takes care of the special case where the constant alloca
represents a variable declared in the condition part of a while loop. In
such a case, this PR removes the constant flag on the alloca operation
when hositing them.