[CIR][CUDA] Handle shared and local variables (#1368)
CUDA shared variables are device-only, accessible from all threads in a
block of some kernel. It's similar to `local` variables in OpenCL which
all threads in a work-group can access. Hence they are realized as
`static` variables in addrspace(local).

On the other hand, the local variables inside a kernel (without special
attributes) are just regular variables, typically emitted by
`CreateTempAlloca`. They are in the default address space.

OG checks if the expected address space, denoted by the type, is the
same as the actual address space indicated by attributes. If they aren't
the same, a `addrspacecast` is emitted when a global variable is
accessed. In CIR however, `cir.get_global` alreadys carries that
information in `!cir.ptr` type, so we don't need a cast.