[CIR] Rename `cir.struct` to `cir.record` and associated changes (#1559)
During the initial upstreaming of `cir.struct` support, a request was
made to rename `cir.struct` to `cir.record` for consistency with its
multiple uses. This change makes the modification along with renaming
various other related classes. I've attempted to also update variable
names and comments to keep everything consistent.

This creates an unfortunate name collision between cir::RecordType and
clang::RecordType, but the impact of that overlaps was relatively small.

The only intended behavioral change is in the names used in the emitted
CIR files and diagnostics.